+++ homebrew/tunara.rb
@@ -1,5 +1,5 @@
 cask "tunara" do
-  version "1.10.0"
+  version "1.11.0"
   sha256 "f2cc8edcdb137e1e8a1fc0cf75e0af76ba0e9fdda13ff843ab39f32e8fc02f0c"
 
   url "https://github.com/24kHandsome1201/tunara/releases/download/v#{version}/Tunara_#{version}_aarch64.dmg"